requirements.txt
streamlit
Pillow
inference-sdk
numpy
streamlit-webrtc
av
opencv-python
